fix: move payment drawer into builder drawer
