+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/FieldListDrawer/field-panels/PaymentsInputPanel.tsx
@@ -7,10 +7,14 @@ import {
   useWatch,
 } from 'react-hook-form'
 import { useDebounce } from 'react-use'
-import { Box, Divider, Flex, FormControl, Text } from '@chakra-ui/react'
+import { Box, FormControl, Text } from '@chakra-ui/react'
 import { cloneDeep } from 'lodash'
 
-import { FormPaymentsField } from '~shared/types'
+import {
+  FormPaymentsField,
+  FormResponseMode,
+  PaymentChannel,
+} from '~shared/types'
 
 import { centsToDollars, dollarsToCents } from '~utils/payments'
 import FormErrorMessage from '~components/FormControl/FormErrorMessage'
@@ -21,21 +25,20 @@ import MoneyInput from '~components/MoneyInput'
 import Toggle from '~components/Toggle'
 
 import { useMutateFormPage } from '~features/admin-form/common/mutations'
+import { useAdminForm } from '~features/admin-form/common/queries'
 
-import { useEnv } from '../../../env/queries'
-import { FormFieldDrawerActions } from '../builder-and-design/BuilderAndDesignDrawer/EditFieldDrawer/edit-fieldtype/common/FormFieldDrawerActions'
-import {
-  setIsDirtySelector,
-  useDirtyFieldStore,
-} from '../builder-and-design/useDirtyFieldStore'
+import { useEnv } from '../../../../../../env/queries'
 import {
   CreatePageDrawerContentContainer,
   useCreatePageSidebar,
-} from '../common'
-import { CreatePageDrawerContainer } from '../common/CreatePageDrawer/CreatePageDrawerContainer'
-import { CreatePageSideBarLayoutProvider } from '../common/CreatePageSideBarLayoutContext'
+} from '../../../../common'
+import { FieldListTabIndex } from '../../../constants'
+import {
+  setIsDirtySelector,
+  useDirtyFieldStore,
+} from '../../../useDirtyFieldStore'
+import { FormFieldDrawerActions } from '../../EditFieldDrawer/edit-fieldtype/common/FormFieldDrawerActions'
 
-import { FormPaymentsDisplay } from './types'
 import {
   dataSelector,
   resetDataSelector,
@@ -51,11 +54,13 @@ const formatCurrency = new Intl.NumberFormat('en-SG', {
   maximumFractionDigits: 2,
 }).format
 
-export const PaymentInput = ({
-  isDisabled,
-}: {
-  isDisabled: boolean
-}): JSX.Element => {
+type FormPaymentsInput = {
+  enabled: boolean
+  description: string
+  display_amount: string
+}
+
+export const PaymentInput = ({ isDisabled }: { isDisabled: boolean }) => {
   const { paymentsMutation } = useMutateFormPage()
 
   const { data: { maxPaymentAmountCents, minPaymentAmountCents } = {} } =
@@ -63,21 +68,18 @@ export const PaymentInput = ({
 
   const setIsDirty = useDirtyFieldStore(setIsDirtySelector)
 
-  const { paymentsData, setData, setToInactive } = usePaymentStore(
-    useCallback(
-      (state) => ({
-        paymentsData: dataSelector(state),
-        setData: setDataSelector(state),
-        setToInactive: setToInactiveSelector(state),
-      }),
-      [],
-    ),
-  )
+  const { paymentsData, setData, setToInactive } = usePaymentStore((state) => ({
+    paymentsData: dataSelector(state),
+    setData: setDataSelector(state),
+    setToInactive: setToInactiveSelector(state),
+  }))
+
+  const { setFieldListTabIndex } = useCreatePageSidebar()
+
+  const handleClose = () => setFieldListTabIndex(FieldListTabIndex.Basic)
 
   // unpack payment data for paymentAmount if it exists
   const paymentAmountCents = paymentsData?.amount_cents
-
-  const { handleClose } = useCreatePageSidebar()
 
   const {
     register,
@@ -85,7 +87,7 @@ export const PaymentInput = ({
     control,
     handleSubmit,
     trigger,
-  } = useForm<FormPaymentsDisplay>({
+  } = useForm<FormPaymentsInput>({
     mode: 'onChange',
     defaultValues: {
       ...paymentsData,
@@ -104,7 +106,7 @@ export const PaymentInput = ({
   }, [dirtyFields, setIsDirty])
 
   const handlePaymentsChanges = useCallback(
-    (paymentsInputs: FormPaymentsDisplay) => {
+    (paymentsInputs: FormPaymentsInput) => {
       const { display_amount, ...rest } = paymentsInputs
       setData({
         ...rest,
@@ -116,7 +118,7 @@ export const PaymentInput = ({
 
   const watchedInputs = useWatch({
     control: control,
-  }) as UnpackNestedValue<FormPaymentsDisplay>
+  }) as UnpackNestedValue<FormPaymentsInput>
 
   const clonedWatchedInputs = useMemo(
     () => cloneDeep(watchedInputs),
@@ -127,53 +129,49 @@ export const PaymentInput = ({
     Object.values(clonedWatchedInputs),
   ])
 
-  const handleCloseDrawer = useCallback(() => handleClose(false), [handleClose])
-
   const paymentIsEnabled = clonedWatchedInputs.enabled
 
-  const amountValidation: RegisterOptions<
-    FormPaymentsDisplay,
-    'display_amount'
-  > = {
-    validate: (val) => {
-      if (!paymentIsEnabled) return true
+  const amountValidation: RegisterOptions<FormPaymentsInput, 'display_amount'> =
+    {
+      validate: (val) => {
+        if (!paymentIsEnabled) return true
 
-      // Validate that it is a money value.
-      // Regex allows leading and trailing spaces, max 2dp
-      const validateMoney = /^\s*(\d+)(\.\d{0,2})?\s*$/.test(val ?? '')
-      if (!validateMoney) return 'Please enter a valid payment amount'
+        // Validate that it is a money value.
+        // Regex allows leading and trailing spaces, max 2dp
+        const validateMoney = /^\s*(\d+)(\.\d{0,2})?\s*$/.test(val ?? '')
+        if (!validateMoney) return 'Please enter a valid payment amount'
 
-      const validateMin =
-        !!minPaymentAmountCents &&
-        !!val &&
-        dollarsToCents(val) >= minPaymentAmountCents
-      // Repeat the check on minPaymentAmountCents for correct typing
-      if (!!minPaymentAmountCents && !validateMin) {
-        return `Please enter a payment amount above ${formatCurrency(
-          Number(centsToDollars(minPaymentAmountCents)),
-        )}`
-      }
+        const validateMin =
+          !!minPaymentAmountCents &&
+          !!val &&
+          dollarsToCents(val) >= minPaymentAmountCents
+        // Repeat the check on minPaymentAmountCents for correct typing
+        if (!!minPaymentAmountCents && !validateMin) {
+          return `Please enter a payment amount above ${formatCurrency(
+            Number(centsToDollars(minPaymentAmountCents)),
+          )}`
+        }
 
-      const validateMax =
-        !!maxPaymentAmountCents &&
-        !!val &&
-        dollarsToCents(val) <= maxPaymentAmountCents
-      // Repeat the check on maxPaymentAmountCents for correct typing
-      if (!!maxPaymentAmountCents && !validateMax) {
-        return `Please enter a payment amount below ${formatCurrency(
-          Number(centsToDollars(maxPaymentAmountCents)),
-        )}`
-      }
-      return true
-    },
-  }
+        const validateMax =
+          !!maxPaymentAmountCents &&
+          !!val &&
+          dollarsToCents(val) <= maxPaymentAmountCents
+        // Repeat the check on maxPaymentAmountCents for correct typing
+        if (!!maxPaymentAmountCents && !validateMax) {
+          return `Please enter a payment amount below ${formatCurrency(
+            Number(centsToDollars(maxPaymentAmountCents)),
+          )}`
+        }
+        return true
+      },
+    }
 
   const handleUpdatePayments = handleSubmit(() => {
     if (isDisabled || !paymentsData) {
       // do not mutate if payments is disabled or unavailable
       return () => {
         setToInactive()
-        handleCloseDrawer()
+        handleClose()
       }
     }
     return paymentsMutation.mutate(
@@ -181,7 +179,7 @@ export const PaymentInput = ({
       {
         onSuccess: () => {
           setToInactive()
-          handleCloseDrawer()
+          handleClose()
         },
       },
     )
@@ -202,6 +200,7 @@ export const PaymentInput = ({
           label="Enable payment"
         />
       </FormControl>
+
       <FormControl
         isReadOnly={paymentsMutation.isLoading}
         isInvalid={!!errors.description}
@@ -219,6 +218,7 @@ export const PaymentInput = ({
         />
         <FormErrorMessage>{errors.description?.message}</FormErrorMessage>
       </FormControl>
+
       <FormControl
         isReadOnly={paymentsMutation.isLoading}
         isInvalid={!!errors.display_amount}
@@ -246,35 +246,26 @@ export const PaymentInput = ({
       <FormFieldDrawerActions
         isLoading={paymentsMutation.isLoading}
         handleClick={handleUpdatePayments}
-        handleCancel={handleCloseDrawer}
-        buttonText="Save payment field"
+        handleCancel={handleClose}
+        buttonText="Save field"
         isDisabled={isDisabled}
-      ></FormFieldDrawerActions>
+      />
     </CreatePageDrawerContentContainer>
   )
 }
 
-type PaymentDrawerProps = {
-  isEncryptMode: boolean
-  isStripeConnected: boolean
-  paymentsField?: FormPaymentsField
-}
+export const PaymentsInputPanel = (): JSX.Element | null => {
+  const { data: form } = useAdminForm()
 
-export const PaymentDrawer = ({
-  isEncryptMode,
-  isStripeConnected,
-  paymentsField,
-}: PaymentDrawerProps): JSX.Element | null => {
-  const { paymentData, setData, resetData } = usePaymentStore(
-    useCallback(
-      (state) => ({
-        paymentData: dataSelector(state),
-        setData: setDataSelector(state),
-        resetData: resetDataSelector(state),
-      }),
-      [],
-    ),
-  )
+  const isEncryptMode = form?.responseMode === FormResponseMode.Encrypt
+  const isStripeConnected =
+    isEncryptMode && form.payments_channel.channel === PaymentChannel.Stripe
+  const paymentsField = isEncryptMode ? form.payments_field : undefined
+
+  const { setData, resetData } = usePaymentStore((state) => ({
+    setData: setDataSelector(state),
+    resetData: resetDataSelector(state),
+  }))
 
   useEffect(() => {
     setData(paymentsField)
@@ -290,31 +281,16 @@ export const PaymentDrawer = ({
   // payment eligibility will be dependent on whether paymentDisabledMessage is non empty
   const isPaymentDisabled = !!paymentDisabledMessage
 
-  // Allows for payment data refresh in encrypt mode
-  if (!paymentData && isEncryptMode) return null
-
   return (
-    <CreatePageSideBarLayoutProvider>
-      <CreatePageDrawerContainer>
-        <Flex pos="relative" h="100%" display="flex" flexDir="column">
-          <Box pt="1rem" px="1.5rem" bg="white">
-            <Flex justify="space-between">
-              <Text textStyle="subhead-3" color="secondary.500" mb="1rem">
-                Edit payment
-              </Text>
-            </Flex>
-            <Divider w="auto" mx="-1.5rem" />
-          </Box>
-          {isPaymentDisabled && (
-            <Box px="1.5rem" pt="2rem" pb="1.5rem">
-              <InlineMessage variant={'info'}>
-                <Text>{paymentDisabledMessage}</Text>
-              </InlineMessage>
-            </Box>
-          )}
-          <PaymentInput isDisabled={isPaymentDisabled} />
-        </Flex>
-      </CreatePageDrawerContainer>
-    </CreatePageSideBarLayoutProvider>
+    <>
+      {isPaymentDisabled && (
+        <Box px="1.5rem" pt="2rem" pb="1.5rem">
+          <InlineMessage variant={'info'}>
+            <Text>{paymentDisabledMessage}</Text>
+          </InlineMessage>
+        </Box>
+      )}
+      <PaymentInput isDisabled={isPaymentDisabled} />
+    </>
   )
 }
