+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/FieldListDrawer/field-panels/index.ts
@@ -1,0 +1,3 @@
+export * from './BasicFieldPanel'
+export * from './MyInfoPanel'
+export * from './PaymentsInputPanel'
