+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/FieldListDrawer/field-panels/usePaymentStore.tsx
@@ -57,10 +57,7 @@ export const usePaymentStore = create<PaymentStore>()(
       if (holding) {
         set({ holdingState: PaymentState.Inactive })
       } else {
-        set({
-          state: PaymentState.Inactive,
-          data: undefined,
-        })
+        set({ state: PaymentState.Inactive })
       }
     },
   })),
