+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignTab.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useState } from 'react'
+import { useCallback, useRef, useState } from 'react'
 import {
   DragDropContext,
   DragStart,
@@ -138,6 +138,8 @@ export const BuilderAndDesignTab = (): JSX.Element => {
 
   const deleteFieldModalDisclosure = useDisclosure()
 
+  const paymentPreviewRef = useRef<HTMLDivElement>(null)
+
   return (
     <DragDropContext
       onDragStart={onDragStart}
@@ -147,6 +149,7 @@ export const BuilderAndDesignTab = (): JSX.Element => {
       <BuilderAndDesignContext.Provider
         value={{
           deleteFieldModalDisclosure,
+          paymentPreviewRef,
         }}
       >
         <BuilderAndDesignDrawer />
