+++ frontend/src/features/admin-form/create/common/CreatePageContent/CreatePageContent.tsx
@@ -6,7 +6,6 @@ import {
 import { BuilderAndDesignTab } from '../../builder-and-design/BuilderAndDesignTab'
 import { EndPageTab } from '../../end-page/EndPageTab'
 import { CreatePageLogicTab } from '../../logic/CreatePageLogicTab'
-import { PaymentTab } from '../../payment/PaymentTab'
 
 export const CreatePageContent = (): JSX.Element => {
   const { activeTab } = useCreatePageSidebar()
@@ -15,8 +14,6 @@ export const CreatePageContent = (): JSX.Element => {
       return <CreatePageLogicTab />
     case DrawerTabs.EndPage:
       return <EndPageTab />
-    case DrawerTabs.Payment:
-      return <PaymentTab />
     default:
       // builder or design
       return <BuilderAndDesignTab />
