+++ frontend/src/features/admin-form/create/common/CreatePageSidebarContext/CreatePageSidebarContext.tsx
@@ -22,16 +22,11 @@ import {
   setToInactiveSelector,
   useFieldBuilderStore,
 } from '../../builder-and-design/useFieldBuilderStore'
-import {
-  setToInactiveSelector as setPaymentToInactiveSelector,
-  usePaymentStore,
-} from '../../payment/usePaymentStore'
 
 export enum DrawerTabs {
   Builder,
   Design,
   Logic,
-  Payment,
   EndPage,
 }
 
@@ -43,7 +38,6 @@ export type CreatePageSidebarContextProps = {
   handleBuilderClick: (shouldBePending: boolean) => void
   handleDesignClick: (shouldBePending: boolean) => void
   handleLogicClick: (shouldBePending: boolean) => void
-  handlePaymentClick: (shouldBePending: boolean) => void
   handleEndpageClick: (shouldBePending: boolean) => void
   handleClose: (shouldBePending: boolean) => void
   isDrawerOpen: boolean
@@ -80,7 +74,6 @@ export const useCreatePageSidebarContext =
     const setFieldsToInactive = useFieldBuilderStore(setToInactiveSelector)
 
     const setDesignState = useDesignStore(setDesignStateSelector)
-    const setPaymentToInactive = usePaymentStore(setPaymentToInactiveSelector)
 
     const [fieldListTabIndex, setFieldListTabIndex] =
       useState<FieldListTabIndex>(FieldListTabIndex.Basic)
@@ -95,10 +88,6 @@ export const useCreatePageSidebarContext =
     useEffect(() => {
       if (activeTab !== DrawerTabs.Design) setDesignState(DesignState.Inactive)
     }, [activeTab, setDesignState])
-
-    useEffect(() => {
-      if (activeTab !== DrawerTabs.Payment) setPaymentToInactive()
-    }, [activeTab, setPaymentToInactive])
 
     const setActiveOrPendingTab = useCallback(
       (tab: DrawerTabs | null, shouldBePending?: boolean) => {
@@ -140,12 +129,6 @@ export const useCreatePageSidebarContext =
       [setActiveOrPendingTab],
     )
 
-    const handlePaymentClick = useCallback(
-      (shouldBePending: boolean) =>
-        setActiveOrPendingTab(DrawerTabs.Payment, shouldBePending),
-      [setActiveOrPendingTab],
-    )
-
     const handleEndpageClick = useCallback(
       (shouldBePending: boolean) =>
         setActiveOrPendingTab(DrawerTabs.EndPage, shouldBePending),
@@ -176,7 +159,6 @@ export const useCreatePageSidebarContext =
       isDrawerOpen,
       handleBuilderClick,
       handleDesignClick,
-      handlePaymentClick,
       handleLogicClick,
       handleEndpageClick,
       handleClose,
