--- frontend/src/features/admin-form/create/payment/types.ts
@@ -1,5 +1,0 @@
-export type FormPaymentsDisplay = {
-  enabled: boolean
-  display_amount?: string
-  description?: string
-}
